# docs/phase6_step2_openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BioVoid Phase 6 Job API",
    "description": "Single-node job orchestration API for controlled Phase 6 productization.",
    "version": "6.0.0-step4"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ready Ready Get"
                }
              }
            }
          }
        }
      }
    },
    "/portal": {
      "get": {
        "summary": "Portal",
        "operationId": "portal_portal_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ops/metrics": {
      "get": {
        "summary": "Ops Metrics",
        "operationId": "ops_metrics_ops_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Ops Metrics Ops Metrics Get"
                }
              }
            }
          }
        }
      }
    },
    "/jobs": {
      "post": {
        "summary": "Submit Job",
        "operationId": "submit_job_jobs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSubmissionResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetailResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/result": {
      "get": {
        "summary": "Download Job Result",
        "operationId": "download_job_result_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "additionalProperties": true,
            "type": "object",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Standard API error envelope."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobDetailResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Utc"
          },
          "started_at_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At Utc"
          },
          "finished_at_utc": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At Utc"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "request": {
            "$ref": "#/components/schemas/JobSubmitRequest"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobErrorResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "created_at_utc",
          "attempts",
          "idempotency_key",
          "request"
        ],
        "title": "JobDetailResponse",
        "description": "Response model for GET /jobs/{job_id}."
      },
      "JobErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "attempts"
        ],
        "title": "JobErrorResponse",
        "description": "Structured job execution failure block."
      },
      "JobInput": {
        "properties": {
          "pdb_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 4,
            "title": "Pdb Id",
            "description": "PDB identifier (e.g. 1CBS)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pdb_id"
        ],
        "title": "JobInput",
        "description": "User-submitted input payload."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed"
        ],
        "title": "JobStatus",
        "description": "Supported job lifecycle statuses."
      },
      "JobSubmissionResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "idempotent_reused": {
            "type": "boolean",
            "title": "Idempotent Reused"
          },
          "created_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Utc"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "idempotent_reused",
          "created_at_utc"
        ],
        "title": "JobSubmissionResponse",
        "description": "Response for POST /jobs."
      },
      "JobSubmitRequest": {
        "properties": {
          "job_type": {
            "type": "string",
            "const": "quick_probe",
            "title": "Job Type",
            "default": "quick_probe"
          },
          "input": {
            "$ref": "#/components/schemas/JobInput"
          },
          "options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input"
        ],
        "title": "JobSubmitRequest",
        "description": "Create-job request model."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}